Make the getting a pixel from an image safe. :children_crossing:
Getting a single pixel from an image was unsafe because a raw pointer has to be dereferenced. There are now checks that ensure that the image pointer was initialized. The actual pixel pointer is reached through pointer arithmetic, so additional checks are performed to make this as safe as possible.

However, the safety of this function still relies on the image being properly initialized by xiGetImage. Any bug there may affect the safety of this function.